+++ python/data/format.py
@@ -17,7 +17,7 @@ class Events:
         assert self.x.dtype == np.uint16
         assert self.y.dtype == np.uint16
         assert self.p.dtype == np.uint8
-        assert self.t.dtype == np.uint32
+        assert self.t.dtype == np.int64
 
         assert self.x.shape == self.y.shape == self.p.shape == self.t.shape
         assert self.x.ndim == 1
@@ -42,7 +42,7 @@ class EventsFloat:
         assert self.x.dtype == np.float32
         assert self.y.dtype == np.float32
         assert self.p.dtype == np.uint8
-        assert self.t.dtype == np.uint32
+        assert self.t.dtype == np.int64
 
         assert self.x.shape == self.y.shape == self.p.shape == self.t.shape
         assert self.x.ndim == 1
@@ -79,7 +79,7 @@ if __name__ == '__main__':
                 x=np.array([0, 1], dtype='uint16'),
                 y=np.array([1, 2], dtype='uint16'),
                 p=np.array([0, 0], dtype='uint8'),
-                t=np.array([0, 5], dtype='uint32')),
+                t=np.array([0, 5], dtype='int64')),
             width=2,
             height=3,
             t_reconstruction=6)
